src/main/java/com/rayan/course/employee/BasePlusCommissionEmployee.java
package com.rayan.course.employee;

/**
 * Created by saeed on 2/2/16.
 */
public class BasePlusCommissionEmployee extends CommissionEmployee{

    private double baseSalary;

    public BasePlusCommissionEmployee(String firstName, String lastName
            , String socialSecurityNumber, double grossSales
            , double commissionRate, double baseSalary) {

        super(firstName, lastName, socialSecurityNumber, grossSales, commissionRate);

        if (baseSalary < 0.0)
            throw new IllegalArgumentException(
                    "Base salary must be >= 0.0");

        this.baseSalary = baseSalary;
    }

    @Override
    public double earnings() {
        return baseSalary + super.earnings();
    }
}
